moved configurable and dyname_class
