+++ src/rail/plotting/plotter.py
@@ -3,9 +3,10 @@ from __future__ import annotations
 import os
 from typing import Any, TYPE_CHECKING
 
+from rail.projects.configurable import Configurable
+from rail.projects.dynamic_class import DynamicClass
+
 from .plot_holder import RailPlotDict
-from .configurable import Configurable
-from .dynamic_class import DynamicClass
 from .validation import validate_inputs
 
 if TYPE_CHECKING:
