+++ src/rail/projects/catalog_template.py
@@ -6,7 +6,7 @@ import itertools
 
 from ceci.config import StageParameter
 
-from rail.plotting.configurable import Configurable
+from .configurable import Configurable
 
 
 class RailProjectCatalogInstance(Configurable):
