+++ src/rail/projects/file_template.py
@@ -6,7 +6,7 @@ import os
 
 from ceci.config import StageParameter
 
-from rail.plotting.configurable import Configurable
+from .configurable import Configurable
 
 
 class RailProjectFileInstance(Configurable):
