+++ src/rail/projects/reducer.py
@@ -11,8 +11,8 @@ from pyarrow import acero
 
 from ceci.config import StageParameter
 
-from rail.plotting.configurable import Configurable
-from rail.plotting.dynamic_class import DynamicClass
+from .configurable import Configurable
+from .dynamic_class import DynamicClass
 
 
 COLUMNS = [
