+++ src/rail/projects/selection_factory.py
@@ -6,7 +6,7 @@ import yaml
 
 
 from ceci.config import StageParameter
-from rail.plotting.configurable import Configurable
+from .configurable import Configurable
 
 
 class RailSelection(Configurable):
